+++ clay/special.clj
@@ -657,27 +657,27 @@
                              :palette gg/palette-blue-1}))]]
 
 (utls/examples-note
-  (special/spherical-bessel-k0 2.3)
-  (special/spherical-bessel-k1 2.3)
-  (special/spherical-bessel-k2 2.3)
-  (special/spherical-bessel-k 3.1 3.2)
-  (special/spherical-bessel-k -3.1 3.2))
+ (special/spherical-bessel-k0 2.3)
+ (special/spherical-bessel-k1 2.3)
+ (special/spherical-bessel-k2 2.3)
+ (special/spherical-bessel-k 3.1 3.2)
+ (special/spherical-bessel-k -3.1 3.2))
 
-;; ## Henkel
+;; ## Hankel
 
 ;; ::: {.callout-tip title="Defined functions"}
-;; * `henkel-1`, `henkel-2`
-;; * `spherical-henkel-1`, `spherical-henkel-2`
+;; * `hankel-1`, `hankel-2`
+;; * `spherical-hankel-1`, `spherical-hankel-2`
 ;; :::
 
-;; Henkel functions of the first and second kind.
+;; Hankel functions of the first and second kind.
 
 ;; $$H^{(1)}_\alpha(x)=J_\alpha(x)+ i Y_\alpha(x)$$
 ;; $$H^{(2)}_\alpha(x)=J_\alpha(x)- i Y_\alpha(x)$$
 
 (utls/examples-note
-  (special/henkel-1 1 2.3)
-  (special/henkel-2 1 2.3))
+ (special/hankel-1 1 2.3)
+ (special/hankel-2 1 2.3))
 
 ;; Spherical functions of the first and second kind.
 
@@ -685,8 +685,8 @@
 ;; $$h^{(2)}_\alpha(x)=j_\alpha(x)- i j_\alpha(x)$$
 
 (utls/examples-note
-  (special/spherical-henkel-1 1 2.3)
-  (special/spherical-henkel-2 1 2.3))
+ (special/spherical-hankel-1 1 2.3)
+ (special/spherical-hankel-2 1 2.3))
 
 ;; ## Erf